Move routes to separate file for clarity
