+++ src/shared/view-types.ts
@@ -1,16 +1,13 @@
 import type {
-  AllResults,
   BenchmarkId,
   CriterionWithoutData,
   ProfileElement,
-  TimelineResponse,
   ValuesPossiblyMissing
 } from './api.js';
 import type {
   CriterionData,
   Environment,
-  RevisionData,
-  Source
+  RevisionData
 } from '../backend/db/types.js';
 import type { ComparisonStatistics, SummaryStatsWithUnit } from './stats.js';
 
@@ -301,54 +298,3 @@ export interface SiteStatsResponse {
   stats: any[];
   version: number;
 }
-
-export const apiRoutes = {
-  '/:projectSlug/source/:sourceId': {
-    method: 'GET',
-    response: undefined as unknown as Source | string
-  },
-  '/rebenchdb/dash/:projectId/results': {
-    method: 'GET',
-    response: undefined as unknown as AllResults[]
-  },
-  '/rebenchdb/dash/:projectId/timeline/:runId': {
-    method: 'GET',
-    response: undefined as unknown as TimelineResponse | null
-  },
-  '/rebenchdb/dash/:projectSlug/profiles/:runId/:commitId': {
-    method: 'GET',
-    response: undefined as unknown as ProfileRow[]
-  },
-  '/rebenchdb/dash/:projectSlug/measurements/:runId/:baseId/:changeId': {
-    method: 'GET',
-    response: undefined as unknown as WarmupDataForTrial[] | null
-  },
-
-  '/rebenchdb/stats': {
-    method: 'GET',
-    response: undefined as unknown as SiteStatsResponse
-  },
-  '/rebenchdb/dash/:projectId/changes': {
-    method: 'GET',
-    response: undefined as unknown as ChangesResponse
-  },
-  '/rebenchdb/dash/:projectId/data-overview': {
-    method: 'GET',
-    response: undefined as unknown as { data: any[] }
-  },
-  '/rebenchdb/dash/:projectName/timelines': {
-    method: 'POST',
-    response: undefined as unknown as TimelineResponse
-  }
-} as const;
-
-interface ApiRoute {
-  method: 'GET' | 'POST';
-  response: unknown;
-}
-
-type Routes = Record<string, ApiRoute>;
-
-const _typeCheckApiRoutes: Routes = apiRoutes;
-
-export type ApiRoutes = typeof apiRoutes;
